+++ pyfms/__init__.py
@@ -1,14 +1,14 @@
 from .cfms import cfms
-from .py_data_override.py_data_override import data_override
-from .py_diag_manager.py_diag_manager import diag_manager
+from .py_data_override import data_override
+from .py_diag_manager import diag_manager
 from .py_field_manager.py_field_manager import FieldTable
-from .py_horiz_interp.py_horiz_interp import horiz_interp
-from .py_mpp.py_mpp import mpp
+from .py_horiz_interp import horiz_interp
+from .py_mpp import mpp
 from .py_mpp.py_mpp_domains import mpp_domains, pyDomain
-from .pyfms import fms
+from .py_fms import fms
 from .utils import data_handling
-from .utils.constants import constants
-from .utils.grid_utils import grid_utils
+from .utils import constants
+from .utils import grid_utils
 
 
 cfms.init()
